fix(@schematics/angular): build in --prod mode by default
According to the idea the CLI should be ready to use, if I want to work in dev, I run `npm start`, and if I want to go in production, I run `npm run build`. But currently the build is in dev mode by default, requiring the user to know he/she has to go to modify the npm script accordingly.