feat(@angular-devkit/schematics): add a workflow class that simplifies the whole thing
Workflows are complete abstraction of the whole system. This makes it easier to
have tools that dont diverge from the standard definition of schematics. In case
tools have to diverge they can implement their own workflow as well.

Please note that workflows are also recursive. This means that you can execute
a schematic inside another schematic. This will happen in the next commit with
the schematic task.

For now, the workflow is relatively crude. It will be improved over then next
few commits.