Create a HostComponent in new component spec
This is how I tend to start every new component I create. I know that this should perhaps not be the default right away, but maybe to be considered at a major version bump.

Thoughts are welcome. Especially since there may be a better strategy for testing components which accept `@Input` of which I am not aware.